Blocks: Move applicable hooks from editor to blocks
