+++ blocks/api/raw-handling/test/index.js
@@ -13,7 +13,7 @@ import { registerCoreBlocks } from '../../../../core-blocks';
 describe( 'rawHandler', () => {
 	beforeAll( () => {
 		// Load all hooks that modify blocks
-		require( 'editor/hooks' );
+		require( 'blocks/hooks' );
 		registerCoreBlocks();
 	} );
