+++ blocks/api/test/parser.js
@@ -43,7 +43,7 @@ describe( 'block parser', () => {
 
 	beforeAll( () => {
 		// Load all hooks that modify blocks
-		require( 'editor/hooks' );
+		require( 'blocks/hooks' );
 	} );
 
 	afterEach( () => {
