+++ blocks/api/test/registration.js
@@ -30,7 +30,7 @@ describe( 'blocks', () => {
 
 	beforeAll( () => {
 		// Load all hooks that modify blocks
-		require( 'editor/hooks' );
+		require( 'blocks/hooks' );
 	} );
 
 	afterEach( () => {
