+++ blocks/hooks/index.js
@@ -4,4 +4,5 @@
 import './align';
 import './anchor';
 import './custom-class-name';
-import './default-autocompleters';
+import './generated-class-name';
+import './layout';
