+++ blocks/hooks/test/generated-class-name.js
