+++ blocks/index.js
@@ -7,4 +7,6 @@
 //
 // Blocks are inferred from the HTML source of a post through a parsing mechanism
 // and then stored as objects in state, from which it is then rendered for editing.
+import './hooks';
+
 export * from './api';
