+++ editor/hooks/align.js
@@ -1,67 +1,19 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
-import { assign, includes } from 'lodash';
+import { includes } from 'lodash';
 
 /**
  * WordPress dependencies
  */
 import { createHigherOrderComponent } from '@wordpress/element';
 import { addFilter } from '@wordpress/hooks';
-import { hasBlockSupport, getBlockSupport } from '@wordpress/blocks';
+import { getBlockValidAlignments } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
 import { BlockControls, BlockAlignmentToolbar } from '../components';
-
-/**
- * Filters registered block settings, extending attributes to include `align`.
- *
- * @param  {Object} settings Original block settings
- * @return {Object}          Filtered block settings
- */
-export function addAttribute( settings ) {
-	if ( hasBlockSupport( settings, 'align' ) ) {
-		// Use Lodash's assign to gracefully handle if attributes are undefined
-		settings.attributes = assign( settings.attributes, {
-			align: {
-				type: 'string',
-			},
-		} );
-	}
-
-	return settings;
-}
-
-/**
- * Returns an array of valid alignments for a block type depending on its
- * defined supports. Returns an empty array if block does not support align.
- *
- * @param  {string}   blockName Block name to check
- * @return {string[]}           Valid alignments for block
- */
-export function getBlockValidAlignments( blockName ) {
-	// Explicitly defined array set of valid alignments
-	const blockAlign = getBlockSupport( blockName, 'align' );
-	if ( Array.isArray( blockAlign ) ) {
-		return blockAlign;
-	}
-
-	const validAlignments = [];
-	if ( true === blockAlign ) {
-		// `true` includes all alignments...
-		validAlignments.push( 'left', 'center', 'right' );
-
-		// ...including wide alignments unless explicitly `false`.
-		if ( hasBlockSupport( blockName, 'wideAlign', true ) ) {
-			validAlignments.push( 'wide', 'full' );
-		}
-	}
-
-	return validAlignments;
-}
 
 /**
  * Override the default edit UI to include new toolbar controls for block
@@ -111,27 +63,6 @@ export const withAlign = createHigherOrderComponent( ( BlockListBlock ) => {
 	};
 }, 'withAlign' );
 
-/**
- * Override props assigned to save component to inject alignment class name if
- * block supports it.
- *
- * @param  {Object} props      Additional props applied to save element
- * @param  {Object} blockType  Block type
- * @param  {Object} attributes Block attributes
- * @return {Object}            Filtered props applied to save element
- */
-export function addAssignedAlign( props, blockType, attributes ) {
-	const { align } = attributes;
-
-	if ( includes( getBlockValidAlignments( blockType ), align ) ) {
-		props.className = classnames( `align${ align }`, props.className );
-	}
-
-	return props;
-}
-
-addFilter( 'blocks.registerBlockType', 'core/align/addAttribute', addAttribute );
 addFilter( 'editor.BlockListBlock', 'core/align/withAlign', withAlign );
 addFilter( 'blocks.BlockEdit', 'core/align/withToolbarControls', withToolbarControls );
-addFilter( 'blocks.getSaveContent.extraProps', 'core/align/addAssignedAlign', addAssignedAlign );
