+++ editor/hooks/anchor.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { assign } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { createHigherOrderComponent, Fragment } from '@wordpress/element';
@@ -23,30 +18,6 @@ import { InspectorAdvancedControls } from '../components';
  * @type {RegExp}
  */
 const ANCHOR_REGEX = /[\s#]/g;
-
-/**
- * Filters registered block settings, extending attributes with anchor using ID
- * of the first node.
- *
- * @param {Object} settings Original block settings.
- *
- * @return {Object} Filtered block settings.
- */
-export function addAttribute( settings ) {
-	if ( hasBlockSupport( settings, 'anchor' ) ) {
-		// Use Lodash's assign to gracefully handle if attributes are undefined
-		settings.attributes = assign( settings.attributes, {
-			anchor: {
-				type: 'string',
-				source: 'attribute',
-				attribute: 'id',
-				selector: '*',
-			},
-		} );
-	}
-
-	return settings;
-}
 
 /**
  * Override the default edit UI to include a new block inspector control for
@@ -84,25 +55,4 @@ export const withInspectorControl = createHigherOrderComponent( ( BlockEdit ) =>
 	};
 }, 'withInspectorControl' );
 
-/**
- * Override props assigned to save component to inject anchor ID, if block
- * supports anchor. This is only applied if the block's save result is an
- * element and not a markup string.
- *
- * @param {Object} extraProps Additional props applied to save element.
- * @param {Object} blockType  Block type.
- * @param {Object} attributes Current block attributes.
- *
- * @return {Object} Filtered props applied to save element.
- */
-export function addSaveProps( extraProps, blockType, attributes ) {
-	if ( hasBlockSupport( blockType, 'anchor' ) ) {
-		extraProps.id = attributes.anchor;
-	}
-
-	return extraProps;
-}
-
-addFilter( 'blocks.registerBlockType', 'core/anchor/attribute', addAttribute );
 addFilter( 'blocks.BlockEdit', 'core/anchor/inspector-control', withInspectorControl );
-addFilter( 'blocks.getSaveContent.extraProps', 'core/anchor/save-props', addSaveProps );
