+++ editor/hooks/custom-class-name.js
@@ -1,10 +1,4 @@
 /**
- * External dependencies
- */
-import { assign } from 'lodash';
-import classnames from 'classnames';
-
-/**
  * WordPress dependencies
  */
 import { createHigherOrderComponent, Fragment } from '@wordpress/element';
@@ -17,27 +11,6 @@ import { hasBlockSupport } from '@wordpress/blocks';
  * Internal dependencies
  */
 import { InspectorAdvancedControls } from '../components';
-
-/**
- * Filters registered block settings, extending attributes with anchor using ID
- * of the first node.
- *
- * @param {Object} settings Original block settings.
- *
- * @return {Object} Filtered block settings.
- */
-export function addAttribute( settings ) {
-	if ( hasBlockSupport( settings, 'customClassName', true ) ) {
-		// Use Lodash's assign to gracefully handle if attributes are undefined
-		settings.attributes = assign( settings.attributes, {
-			className: {
-				type: 'string',
-			},
-		} );
-	}
-
-	return settings;
-}
 
 /**
  * Override the default edit UI to include a new block inspector control for
@@ -74,25 +47,4 @@ export const withInspectorControl = createHigherOrderComponent( ( BlockEdit ) =>
 	};
 }, 'withInspectorControl' );
 
-/**
- * Override props assigned to save component to inject anchor ID, if block
- * supports anchor. This is only applied if the block's save result is an
- * element and not a markup string.
- *
- * @param {Object} extraProps Additional props applied to save element.
- * @param {Object} blockType  Block type.
- * @param {Object} attributes Current block attributes.
- *
- * @return {Object} Filtered props applied to save element.
- */
-export function addSaveProps( extraProps, blockType, attributes ) {
-	if ( hasBlockSupport( blockType, 'customClassName', true ) && attributes.className ) {
-		extraProps.className = classnames( extraProps.className, attributes.className );
-	}
-
-	return extraProps;
-}
-
-addFilter( 'blocks.registerBlockType', 'core/custom-class-name/attribute', addAttribute );
 addFilter( 'blocks.BlockEdit', 'core/custom-class-name/inspector-control', withInspectorControl );
-addFilter( 'blocks.getSaveContent.extraProps', 'core/custom-class-name/save-props', addSaveProps );
